refactor: migrate redis cli
